Flashes ID are now consistently starting from 1
However, to maintain links to metadata files, which are already public and cannot be changed, the names of the metadata files are not changed.
Instead, there's a new column in the database "flashes" table, which stores the old flash ID and now the name of the metadata file.
New flashes will have their metadata labeled with the same ID as the Flash